Update README.md
The accompanying code repositories have been meticulously developed to fulfill the requirements of the course assignments associated with Statistical Techniques in Physics. The file designated as Statistiki_Ergasia_1 encompasses the programming solutions pertinent to the first assignment. Similarly, the files named Statistiki_Ergasia_2 and Statistiki_Ergasia_3 encapsulate the code implementations for the second and third assignments, respectively. Each file serves as a comprehensive solution set, adhering to the specified assignment guidelines and demonstrating robust application of statistical methodologies within the realm of physics.