Update README.md
The accompanying code repositories have been meticulously developed to fulfill the requirements of the course assignments associated with Statistical Techniques in Physics. The file designated as Statistiki_Ergasia_1 encompasses the programming solutions pertinent to the first assignment. Similarly, the files named Statistiki_Ergasia_2 and Statistiki_Ergasia_3 encapsulate the code implementations for the second and third assignments, respectively. Additionally, a significant portion of the code across all assignments has been seamlessly integrated and presented within a Jupyter Notebook, utilizing the Python programming language. This integration not only facilitates a more interactive and user-friendly exploration of the code but also demonstrates a comprehensive application of statistical methodologies within the realm of physics. Each file serves as a complete solution set, adhering to the specified assignment guidelines and showcasing a robust application of statistical and computational techniques.